requirements.txt
pyupbit==0.2.34
python-dotenv==1.0.0
requests==2.31.0
google-generativeai>=0.4.0
websocket-client==1.6.4

